+++ src/layouts/MainLayout.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Outlet } from 'react-router-dom'
+
+export const MainLayout: React.FC = () => {
+  return (
+    <div>
+      <Outlet />
+    </div>
+  )
+}
