+++ src/pages/Welcome1.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { NavLink } from 'react-router-dom'
+
+export const Welcome1: React.FC = () => {
+  return (
+    <div>
+      1<NavLink to='/welcome/2'>下一页</NavLink>
+    </div>
+  )
+}
