+++ src/pages/Welcome2.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { NavLink } from 'react-router-dom'
+
+export const Welcome2: React.FC = () => {
+  return (
+    <div>
+      2<NavLink to='/welcome/3'>下一页</NavLink>
+    </div>
+  )
+}
