+++ src/pages/Welcome3.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { NavLink } from 'react-router-dom'
+
+export const Welcome3: React.FC = () => {
+  return (
+    <div>
+      3<NavLink to='/welcome/4'>下一页</NavLink>
+    </div>
+  )
+}
