+++ src/pages/Welcome4.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { NavLink } from 'react-router-dom'
+
+export const Welcome4: React.FC = () => {
+  return (
+    <div>
+      <div>
+        4<NavLink to='/welcome/xxxx'>下一页</NavLink>
+      </div>
+    </div>
+  )
+}
