--- src/pages/layouts/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Index() {
-  return (
-    <div>
-      layouts
-    </div>
-  )
-}
